+++ fiblary/__init__.py
@@ -18,13 +18,4 @@ API implementation for Home Center 2
 
 """
 
-__all__ = ['__version__']
-
-import pbr.version
-
-version_info = pbr.version.VersionInfo('fiblary')
-
-try:
-    __version__ = version_info.version_string()
-except AttributeError:
-    __version__ = None
+__version__ = '0.1.2'