+++ ChangeLog.md
@@ -1,3 +1,7 @@
+## 0.1.2
+
+  - Simplified the installation.
+
 ## 0.1.1
 
 Bugfixes:
@@ -6,7 +10,7 @@ Bugfixes:
 
 Features:
 
-  - Event manager added to the clinet class with example code
+  - Event manager added to the client class with example code
 
 ## 0.1.0